Define Macro EXPECT_TIXI_ERROR for unit testing + minor changes
This commit defines the macro EXPECT_TIXI_ERROR to see if a command throws an exception with the expected  return code. It also changes the install path for the c++ bindings to the default include directory. Finally it renames the macro definition HAVE_CXX11 to TIXI_HAVE_CXX11 to make sure it doesn't mess with the tigl macro of the same name

fixes #82